Moving manager services
