+++ backend/src/services/Manager/AuthenticateManagerService.ts
